# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock-movements:
    get:
      operationId: listStockMovements
      tags: [inventory]
      summary: List stock movements
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [inventory]
      summary: Record a stock in/out movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Created movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"

  # Reports
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  # End of Day
  /end-of-day:
    get:
      operationId: listEndOfDay
      tags: [endofday]
      summary: List end of day records
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of EOD records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EndOfDay"
    post:
      operationId: closeDay
      tags: [endofday]
      summary: Close the day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseDayInput"
      responses:
        "201":
          description: Day closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndOfDay"
  /end-of-day/preview:
    get:
      operationId: getEndOfDayPreview
      tags: [endofday]
      summary: Preview end of day summary for a date
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: EOD preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndOfDayPreview"

  # Inventory Loss
  /inventory-loss:
    get:
      operationId: listInventoryLoss
      tags: [inventoryloss]
      summary: List inventory loss records
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of inventory loss records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryLoss"
    post:
      operationId: createInventoryLoss
      tags: [inventoryloss]
      summary: Record an inventory loss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryLossInput"
      responses:
        "201":
          description: Created inventory loss record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryLoss"
components:
  schemas:
...
    MessageResult:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    StockMovement:
      type: object
      required: [id, productId, type, quantity, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: integer
        reason:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
    StockMovementInput:
      type: object
      required: [productId, type, quantity]
      properties:
        productId:
          type: integer
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: integer
        reason:
          type: string

    # Dashboard
...
    Document:
      type: object
      required: [id, title, category, fileName, filePath, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        fileName:
          type: string
        filePath:
          type: string
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string

    # End of Day
    EndOfDay:
      type: object
      required: [id, date, totalCashSales, totalWalletSales, totalBankSales, totalRevenue, totalProfit, totalTransactions, closedAt]
      properties:
        id:
          type: integer
        date:
          type: string
        totalCashSales:
          type: number
        totalWalletSales:
          type: number
        totalBankSales:
          type: number
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalTransactions:
          type: integer
        notes:
          type: ["string", "null"]
        closedBy:
          type: ["integer", "null"]
        closedByName:
          type: ["string", "null"]
        closedAt:
          type: string
        createdAt:
          type: string
    EndOfDayPreview:
      type: object
      required: [date, totalCashSales, totalWalletSales, totalBankSales, totalRevenue, totalProfit, totalTransactions, alreadyClosed]
      properties:
        date:
          type: string
        totalCashSales:
          type: number
        totalWalletSales:
          type: number
        totalBankSales:
          type: number
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalTransactions:
          type: integer
        alreadyClosed:
          type: boolean
    CloseDayInput:
      type: object
      required: [date]
      properties:
        date:
          type: string
        notes:
          type: string

    # Inventory Loss
    InventoryLoss:
      type: object
      required: [id, type, productId, quantity, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [return, damaged, lost]
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        reason:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
    InventoryLossInput:
      type: object
      required: [type, productId, quantity]
      properties:
        type:
          type: string
          enum: [return, damaged, lost]
        productId:
          type: integer
        quantity:
          type: integer
        reason:
          type: string