# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  fileName: string;
  filePath: string;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  createdBy?: number | null;
  /** @nullable */
  createdByName?: string | null;
  createdAt: string;
  updatedAt?: string;
}