# lib/api-zod/src/generated/types/endOfDay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface EndOfDay {
  id: number;
  date: string;
  totalCashSales: number;
  totalWalletSales: number;
  totalBankSales: number;
  totalRevenue: number;
  totalProfit: number;
  totalTransactions: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  closedBy?: number | null;
  /** @nullable */
  closedByName?: string | null;
  closedAt: string;
  createdAt?: string;
}